Adding LicenseService dependency injection
Building on the request of @pgwillia on #1047, I'm exposing a means of
leveraging a different license service class. This is useful as some
places allow the user to free-form fill out a license; While I'm not
sure of the general usage, this is a demonstration of how it might be
done.